Extract SavePlay and SharePlay
