+++ src/utils/locales/en.js
@@ -115,6 +115,7 @@ export default {
     saveModificationsTitle: 'Do you want to save the current play?',
     saveModificationsText:
       "The play '{{title}}' has been modified.\n\nDo you want to save these modifications before closing it?",
+    saveSuccess: "The play has been saved as '{{title}}'",
     savedPlaysList: {
       title: 'Saved plays',
       empty: "You haven't saved any plays yet",
@@ -122,20 +123,17 @@ export default {
       deleteConfirmation: 'Do you really want to delete this play?',
       deleteSuccess: "The play '{{title}}' has been deleted.",
     },
-    save: 'Save',
-    saveSuccess: "The play has been saved as '{{title}}'",
-    rename: 'Rename',
-    new: 'New',
-    share: 'Share',
-    shareTitle: 'Sharing {{title}}',
-    shareMessage: 'Use the following link to import my awesome play in Disc In: {{url}}',
-    shareError: 'It looks like there was an error sharing your play',
     renamePlayModal: {
       placeholder: 'Click here to enter the new name',
       alreadyExists: 'This name already exists',
       empty: 'You cannot set an empty name',
       renameSuccess: 'The play was successfully renamed!',
       cta: 'Apply',
+    },
+    sharePlay: {
+      shareTitle: 'Sharing {{title}}',
+      shareMessage: 'Use the following link to import my awesome play in Disc In: {{url}}',
+      shareError: 'It looks like there was an error sharing your play',
     },
   },
   drills: {
